+++ puncturedfem/antilap/antilap.py
@@ -1,144 +1,126 @@
 import numpy as np
-from scipy.sparse.linalg import gmres, LinearOperator
-
-from .. import mesh
-from .. import d2n
+from ..mesh.cell import cell
+from ..d2n.fft_deriv import fft_antiderivative
+from ..d2n.log_terms import get_log_grad
+from . import log_antilap
 from .. import nystrom
 
-def get_anti_laplacian_harmonic(K: mesh.cell, psi, psi_hat, debug=False):
+def get_anti_laplacian_harmonic(K: cell, psi, psi_hat, a):
+	"""
+	Returns the trace and weighted normal derivative of an anti-Laplacian of a
+	harmonic function
+		phi = psi + sum_{j=1}^m a_j ln |x - xi_j|
+	given the trace of psi, the trace of its harmonic conjugate psi_hat,
+	and the logarithmic coefficients a_1,...,a_m
+	(When K is simply connected, phi = psi and a is an empty list)
+	"""
 
 	if K.num_holes == 0:
-		# simply-connected
-		PSI, dPSI_dn_wgt = _get_anti_laplacian_simply_connected(
-			K, psi, psi_hat
-		)
+		PHI, PHI_wnd = _antilap_simply_connected(K, psi, psi_hat)
 	else:
-		# multiply-connected
-		PSI, dPSI_dn_wgt = _get_anti_laplacian_multiply_connected(
-			K, psi, psi_hat, debug
-		)
+		PHI, PHI_wnd = _antilap_multiply_connected(K, psi, psi_hat, a)
 
-	return PSI, dPSI_dn_wgt
+	return PHI, PHI_wnd
 
-def _get_anti_laplacian_simply_connected(K: mesh.cell, psi, psi_hat):
+def _antilap_simply_connected(K: cell, phi, phi_hat):
 
-	rho = _get_rho_from_grad(K, psi, - psi_hat)
-	rho_hat = _get_rho_from_grad(K, psi_hat, psi)
-
-	x1, x2 = K.get_boundary_points()
-
-	PSI = 0.25 * (x1 * rho + x2 * rho_hat)
-
-	PSI_x1 = 0.25 * (rho 	 + x1 * psi + x2 * psi_hat)
-	PSI_x2 = 0.25 * (rho_hat + x2 * psi - x1 * psi_hat)
-
-	dPSI_dn = K.dot_with_normal(PSI_x1, PSI_x2)
-	dPSI_dn_wgt = K.multiply_by_dx_norm(dPSI_dn)
-
-	return PSI, dPSI_dn_wgt
-
-def _get_rho_from_grad(K, g1, g2):
+	# length of interval of integration in parameter space
 	interval_length = 2 * np.pi * K.num_edges
-	drho_dt = K.dot_with_tangent(g1, g2)
-	drho_dt_wgt = K.multiply_by_dx_norm(drho_dt)
-	return d2n.fft_deriv.fft_antiderivative(drho_dt_wgt, interval_length)
 
-def _get_anti_laplacian_multiply_connected(K: mesh.cell, psi, psi_hat, debug):
+	# integrate tangential derivative of rho
+	rho_td = K.dot_with_tangent(phi, -phi_hat)
+	rho_wtd = K.multiply_by_dx_norm(rho_td)
+	rho = fft_antiderivative(rho_wtd, interval_length)
 
-	# single and double layer operators
-	T1 = nystrom.single_layer.single_layer_mat(K)
-	T2 = nystrom.double_layer.double_layer_mat(K)
+	# integrate tangential derivative of rho_hat
+	rho_hat_td = K.dot_with_tangent(phi_hat, phi)
+	rho_hat_wtd = K.multiply_by_dx_norm(rho_hat_td)
+	rho_hat = fft_antiderivative(rho_hat_wtd, interval_length)
 
-	# get points on the boundary
+	# coordinates of boundary points
 	x1, x2 = K.get_boundary_points()
 
-	# shift the origin to a point outside of K
-	x1 -= K.ext_pt[0]
-	x2 -= K.ext_pt[1]
+	# construct anti-Laplacian
+	PHI = 0.25 * (x1 * rho + x2 * rho_hat)
 
-	# square norm of boundary points in shifted coordinates
-	x_norm_sq = x1 ** 2 + x2 ** 2
+	# gradient of anti-Laplacian
+	PHI_x1 = 0.25 * (rho 	 + x1 * phi + x2 * phi_hat)
+	PHI_x2 = 0.25 * (rho_hat + x2 * phi - x1 * phi_hat)
 
-	# shifted boundary points dotted with unit tangent and normal
-	xt = K.dot_with_tangent(x1, x2)
-	xn = K.dot_with_normal(x1, x2)
+	# weigthed normal derivative of anti-Laplacian
+	PHI_nd = K.dot_with_normal(PHI_x1, PHI_x2)
+	PHI_wnd = K.multiply_by_dx_norm(PHI_nd)
 
-	# rescale by square norm
-	xt /= x_norm_sq
-	xn /= x_norm_sq
+	return PHI, PHI_wnd
 
-	# mulitpy by derivative norm in prepartion for integration
-	xt = K.multiply_by_dx_norm(xt)
-	xn = K.multiply_by_dx_norm(xn)
+def _antilap_multiply_connected(K: cell, psi, psi_hat, a):
 
-	# array sizes
-	N = K.num_pts
+	# compute F * t and \hat F * t
+	F_t = K.dot_with_tangent(psi, -psi_hat)
+	F_hat_t = K.dot_with_tangent(psi_hat, psi)
 
-	# row sums of double layer operator
-	T2_sum = np.sum(T2, 1)
+	# compute b_j and c_j
+	b = np.zeros((K.num_holes,))
+	c = np.zeros((K.num_holes,))
+	for j in range(K.num_holes):
+		b[j] = K.integrate_over_specific_contour(F_hat_t, j + 1) / (-2 * np.pi)
+		c[j] = K.integrate_over_specific_contour(F_t, j + 1) / (2 * np.pi)
 
-	integrator_mat = d2n.harmconj.get_integrator(K)
+	# compute \mu_j and \hat\mu_j
+	mu, mu_hat = get_log_grad(K)
+	mu_hat *= -1
 
-	# block RHS
-	b = np.zeros((2 * N,))
-	b[:N] = T1 @ (+ xn * psi + xt * psi_hat)
-	b[N:] = T1 @ (- xt * psi + xn * psi_hat)
+	# compute \psi_0 and \hat\psi_0
+	psi_0 = psi - (mu @ b - mu_hat @ c)
+	psi_hat_0 = psi_hat - (mu @ c + mu_hat @ b)
 
-	# define linear operator function
-	def linop4antilap(x):
+	# compute weighted normal derivatives of rho and rho_hat
+	rho_nd_0 = K.dot_with_normal(psi_0, -psi_hat_0)
+	rho_wnd_0 = K.multiply_by_dx_norm(rho_nd_0)
+	rho_hat_nd_0 = K.dot_with_normal(psi_hat_0, psi_0)
+	rho_hat_wnd_0 = K.multiply_by_dx_norm(rho_hat_nd_0)
 
-		# unpack
-		rho = x[:N]
-		rho_hat = x[N:]
+	# solve for rho_0 and rho_hat_0
+	rho_0 = nystrom.neumann.solve_neumann_zero_average(K, rho_wnd_0)
+	rho_hat_0 = nystrom.neumann.solve_neumann_zero_average(K, rho_hat_wnd_0)
 
-		# allocate space
-		y = np.zeros((2 * N,))
+	# compute anti-Laplacian of psi_0
+	x1, x2 = K.get_boundary_points()
+	PHI = 0.25 * (x1 * rho_0 + x2 * rho_hat_0)
+	PHI_x1 = 0.25 * (rho_0     + x1 * psi_0 + x2 * psi_hat_0)
+	PHI_x2 = 0.25 * (rho_hat_0 + x2 * psi_0 - x1 * psi_hat_0)
+	PHI_nd = K.dot_with_normal(PHI_x1, PHI_x2)
+	PHI_wnd = K.multiply_by_dx_norm(PHI_nd)
 
-		# first block row
-		y[:N] = d2n.harmconj.apply_T2(
-			rho, T2, T2_sum, K.closest_vert_idx
-		)
-		y[:N] += T1 @ (+ xn * rho + xt * rho_hat)
-		y[:N] += integrator_mat @ rho
+	# compute M = sum_j M_j
+	for j in range(K.num_holes):
+		xi = K.hole_int_pts[:, j]
+		x_xi_1 = x1 - xi[0]
+		x_xi_2 = x2 - xi[1]
+		x_xi_norm_sq = x_xi_1 ** 2 + x_xi_2 ** 2
+		log_x_xi_norm = 0.5 * np.log(x_xi_norm_sq)
+		PHI += 0.5 * (b[j] * x_xi_1 + c[j] * x_xi_2) * log_x_xi_norm
+		M_x1 = 0.5 * (b[j] * mu[:, j] - c[j] * mu_hat[:, j]) *x_xi_1 \
+			+ 0.5 * b[j] * log_x_xi_norm
+		M_x2 = 0.5 * (b[j] * mu[:, j] - c[j] * mu_hat[:, j]) *x_xi_2 \
+			+ 0.5 * c[j] * log_x_xi_norm
+		M_nd = K.dot_with_normal(M_x1, M_x2)
+		PHI_wnd += K.multiply_by_dx_norm(M_nd)
 
-		# second block row
-		y[N:] = d2n.harmconj.apply_T2(
-			rho_hat, T2, T2_sum, K.closest_vert_idx
-		)
-		y[N:] += T1 @ (- xt * rho + xn * rho_hat)
-		y[N:] += integrator_mat @ rho_hat
+	# compute \Lambda_j
+	PHI += log_antilap.get_log_antilap(K) @ a
+	PHI_wnd += log_antilap.get_log_antilap_weighted_normal_derivative(K) @ a
 
-		return y
+	return PHI, PHI_wnd
 
-	# define linear operator object
-	A = LinearOperator(
-		dtype = float,
-		shape = (2 * N, 2 * N),
-		matvec = linop4antilap
-	)
+def _rational_function_coefficients(K: cell, F_t, F_hat_t):
+	# TODO
+	pass
 
-	# solve Nystrom system
-	x, flag = gmres(A, b, atol=1e-12, tol=1e-12)
+def _antilap_rational_terms(K:cell, b, c):
+	# TODO
+	pass
 
-	# unpack solution components
-	rho = x[:N]
-	rho_hat = x[N:]
-
-	# weighted normal derivative of rho
-	drho_dn_wgt = xn * (psi - rho) + xt * (psi_hat - rho_hat)
-
-	# trace of anti-Laplacian
-	PSI = 0.25 * x_norm_sq * rho
-
-	# weighted normal derivative of anti-Laplacian
-	dPSI_nd_wgt = x_norm_sq * (0.5 * xn * rho + 0.25 * drho_dn_wgt)
-
-	# DEBUG
-	if debug:
-		print(f'gmres flag = {flag}')
-		print('residual = %.4e'%(np.linalg.norm(A(x) - b)))
-		d2n.harmconj.get_cond_num(A)
-		print(integrator_mat @ rho)
-		print(integrator_mat @ rho_hat)
-
-	return PSI, dPSI_nd_wgt
+def _antilap_log_terms(K: cell):
+	# TODO
+	pass