+++ puncturedfem/locfun/__init__.py
@@ -1,3 +1,1 @@
 from . import locfun
-from . import almost_divide
-from . import unweight