+++ puncturedfem/nystrom/__init__.py
@@ -1,2 +1,4 @@
 from . import single_layer
 from . import double_layer
+from . import neumann
+from . import apply_double_layer